Adapts haskanoid to run using BearRiver.
This commit modifies haskanoid to run using the current implementation of
BearRiver (an FRP library that implements a subset of Yampa, built on top of
Dunai, a monadic reactive library).

There are two adaptations:

- The paddle position is updated immediately (not based on the velocity).  This
  is because the current implementation of the ArrowLoop instance in BearRiver
actually loops (triggering an exception that the RTS catches and reports as
<<loop>>). This limitation should be address and this patch unapplied.

- The cabal file introduces a new flag, bearriver, disabled by default,
that uses bearriver instead of Yampa to compile haskanoid. BearRiver uses
the same module namespace to make compilation easier.