Parameterise BearRiver over a monad to enable Monadic Classic FRP.
The current version of BearRiver fixes the monad to Identity
in its definition of ClockInfo. This gets in the way when we
want to use impure monads.

This commit parameterises the definition over a monad, but makes
FRP.Yampa (the actual Yampa replacement) re-define SF to a pure version
(BearRiver's SF applied to the Identity monad).