Renames module Data.VectorSpace.Specific.
The module Data.VectorSpace.Specific contained instances for specific Haskell
types. Examples of such types include Int and Double.

This was (to me) a bit unintuitive, as I would consider a type specific,
without a clear, well-established meaning for that term, maybe if it is not
polymorphic.

This commit renames Data.VectorSpace.Specific to Data.VectorSpace.Primitive.
This carries the more intuitive (imho) idea that these types are primitive
Haskell types, that is, those that are not formed of simpler type definitions
using the type-level language available in Haskell.

Closes #46.