Moves wormholes into examples.
Dunai is getting too big. We have a re-structuring coming up.
The more we add, the harder it is to work with.

This fits perfectly in the examples, so let's temporarily put
it there. Other very similar libraries (like the implementation
of Keera Hails on top of Dunai) share features with wormholes
and they are also there.

Later we'll discuss what should be in Dunai major version +2.