Disables the combination of discrete with GHC 7.8.4.
Travis currently fails to compile all examples when this combination is
enabled. This may be a consequence of installing wx, which may depend on
newer versions of libraries that do not expose certain modules.

I have tried it on my machine with GHC 7.8.4 and it works. What's
strange is: if you try to install discrete, it won't install
transformers, which is a direct dependency of dunai and it includes a
missing module that triggers the error.

From this point on, I don't fully understand why it won't work on
travis.

Unfortunately, I don't have time to investigate too deeply. Welcome to
the world of imperfect engineering practices due to limited resources :(