Relax Type-Class Constraints
The implementation of many functions and type-class instances had
a constraints on `m` for it to have an instance of `Monad`.
In some cases, this was brought in by functions in the InternalCore,
such as `morphGS` (to apply a Natural Transformation) and `feedback`.

These functions, however, were only using a single bind followed by
a `return`, which monad laws make equal to `fmap`. Based on this, we
reimplement this methods from a `do` notation to an `fmap`, and thus
replace a `Monad` constraint with a `Functor` constraint.