Implement afterEach and afterEachCat. Refs #148.
The functions afterEach and afterEachCat were never implemented
in bearriver.

This introduces two implementations. The one for afterEach is just a
copy of the one for Yampa, with a different signature. The one for
afterEachCat is re-implemented from Yampa by decoupling processing the
queue of events (with span) and providing the continuations.