Deprecate dunai's VectorSpace in favour of simple-affine-space.
Refs #172, #13.

The Data.VectorSpace module in dunai conflicts with the one in
simple-affine-space, which is used by Yampa. Ideally, bearriver would
use the same library, and apps written for Yampa would work out of the
box. This was not the case.

This change removes the module Data.VectorSpace from dunai completely,
and uses the definitions in simple-affine-space instead. As a
consequence of that:

- RModule no longer exists.
- Dunai needs a major version bump, and a new release.
- Bearriver needs a constraint on dunai's version number and a new release.
- Bearriver re-exports Data.VectorSpace from simple-affine-space, just
like Yampa does.