Update code in Section 6.1 of paper to compile. Refs #96.
The code example in reactiveprogramming.lhs comes directly from the
paper. There are a number of changes needed to compile it as dunai
stands today. Some are motivated by updates to dunai itself, but some
are the result of either small errors in the paper or just liberties
taken to facilitate reading.

This change introduces the following changes to that module:
- It enables extentions needed to compile the module. In particular,
we need extensions to be able to define the type Sink and the
Contravariant instance.

- It imports modules needed to compile the module. In particular, we
need MSFs, Contravariant, TypeCompose and the Num instances for MSFs.

- Sink is defined as a type but then partially applied, which is not
allowed. We can avoid this by using TypeCompose's Flip newtype, and
unFlipping sinks where needed. This is obviously not ideal, but for the
sake of example, it works.

- The Contravariant instance of Sink includes the signature of
contramap, which requires the GHC extension InstanceSigs.

- The contravariant instance of Sink does not include the Monad
constraint on m, which is ok to make the paper easier to read, but
simply won't compile.

- mirroredMouseX is defined twice, so the second definition is ignored.

- mouseX is required, so a type-correct definition (undefined) is
provided.