Include links to four talks on MSFs and FRP. Refs #175.
There are a number of talks on MSFs and FRP (using MSFs as an
implementation vehicle). This commit adds links to those talks to the
readme, to the existing videos section.

I clicked on one of those and heard myself speak in public while doing
this. Let the record reflect for all posterity that this is the most
painful commit I've ever had to submit, and I died inside a little while
hearing myself talk. No amount of puppies will ever make up for my pain.
Ok, maybe 3.