Remove extra time step during switches. Refs #180.
In Yampa, signal functions are started at time 0 (zero). This
is true at the very beginning of the simulation and after each
switch.

Bearriver's switch contains an extra step. The signal function
switched on after a switch is started with a non-zero time delta,
introducing a time jump.

This change eliminates that extra step.

This change also makes it clearer that reactimate is using dunai's
switch (not bearriver's), and that the conversion to Event and back to
Maybe is unnecessary because reactimate is implemented at the level of
abstraction of Dunai's MSFs, not Bearriver's SFs.

Note: commit manually modified by Ivan Perez <ivan.perez@keera.co.uk> to
rebase it on top of 'develop', to document the change and refer to the
issue number. Original PR: #177.