Compile with GHC 8.6.5 and do not allow failures. Refs #178.
Version 8.6 of GHC has been around for quite some time and it is
now stable.

This change modifies the travis file to use GHC 8.6.5 instead of
GHC 8.6.1 (the latest version without bugs at the time when it
was last updated), and to not allow failures in that branch
(otherwise, we may not realize that we broke something).