Add examples from Sec7 and compile the ones already present. Refs #97.
This commit fixes the bugs and updates the code in the existing frp.lhs
file from the original FRP Refactored paper, and adds a separate module
with a basic FRP implementation following the description in the paper.

To be as close a possible, it needs of a GHC extension. Also, a mistake
in a signature has been corrected.