Avoid file name clash on OSX. Refs #182.
OSX does not distinguish between uppercase and lowercase file names. As
a consequence, the files tests/hlint.hs and tests/HLint.hs cannot
co-exist. This breaks the tests, and is particularly problematic
when cloning the repo, as it is never clean (git always reports
modifications), which limits certain git operations (some operations
cannot be applied to repos with modifications).

This change renames the actual test program to HLintMain, in accordance
with the decision made in other repos I worked on. The corresponding
test is changed in the cabal file.