Adjust maintainer to point to me. Refs #186.
This file comes from the file hlint.hs, borrowed from the lens
projects, created by Edward Kmett. The copyright in the file
mentions Edward, and the maintainer was him as well.

This change modifies the maintainer to point to me, keeping the (c) line
the same, understanding that the copyright for this file is his.