Depend on transformers-compat if GHC 7.8 is used. Refs #185.
The build matrix of dunai on hackage is reporting a problem when
trying to build with GHC 7.8.

Although travis is not reporting an issue, compiling locally indicates
that, with the latest list of packages available on hackage, cabal finds
a build plan where one of the modules required by dunai is hidden
(more specifically, is in a package that is hidden).

This change modifies the dependencies of dunai so that that package is
available, but only for older versions of GHC (7.8 and prior).