Completes undefined functions. Refs #95.
This change completes several functions left as undefined, and adds
necessary constraints to signatures (mainly MonadPlus). The left-variant
definitions added by this commit are copy-and-paste-and-replace of
right-variant functions with analogous names.

The function andThen, mentioned in the paper, is given a more
constrained signature so that the type-checker can figure out that the
monad is the same all across. The signature is copied from the paper.