Remove Haskanoid. Refs #190.
The game Haskanoid now officially supports bearriver with a flag in the
cabal file. That makes it unnecessary to keep this version around in the
dunai repo.

This change eliminates the haskanoid example altogether. Because it is
not named in the travis file, that does not need to change. The change
is documented in the readme, to point to the repo instead of a folder
in examples/.