Avoid duplicate effects in LTL evaluation. Refs #195.
The existing LTL evaluation engine performs the same effect at the same
sample twice. This is because the evaluation of the current sample
happens both to determine the current boolean value of the temporal
predicate, and to drop the head of the stream.

This change implements LTL evaluation without that duplication of
effects. It internally uses a multi-valued logic and later drops into a
bi-valued logic once the end of the stream is reached or a decision can
be made.