Fix leak in Bearriver's reactimate. Refs #194.
The current implementation of reactimate in bearriver has a leak.
This has been reported by @walseb, and correctly exemplified.
In particular, a simple reactimate call like:

reactimate (return NoEvent)
           (\ _ -> return (0.1, Nothing))
           (\_ _ -> pure False)
           (constant ())

currently creates a memory leak. Internally, the reason behind the leak
is that the output of keepLast is not necessarily evaluated, and is kept
in memory.

This commit forces WHNF evaluation of the partial result of keepLast in
FRP.BearRiver.reactimate, to avoid that leak. It has been validated by
running the example above, which now produces a flatter profile.

Thank you to @walseb for reporting the leak, showing evidence with the
memory profile and heap profile, and providing a minimal reproducible
example. Thank you to @turion for pointing to the implementation of
keepLast that reactimate is based on, and proposing the solution finally
accepted into bearriver.