Fix classicfrp example and put it on CI. Refs #214.
This repo includes an example that is not included in the CI
infrastructure. The example was built using dunai<=0.4, and used an old
interface.

One of the functions (liftMSFPurer) was eliminated as of that version,
and renamed to morphS. Because the example was not in the CI, it was not
updated and the error slipped through.

This commit performs two changes:
- Update the example to compile correctly.
- Add it to the CI, so it is always checked.

This commit has been checked through multiple travis builds, and is the
result of rebasing on top of other PRs, so CI is disabled on purpose.

[ci skip]