Add dunai-test to CI. Refs #214.
Dunai test is considered an integral part of the offering of dunai.  At
present, it is not added to the build, so it breaks often and is not
being maintained with the same level of quality one could expect from
dunai.

This commit adds dunai test to the travis build. It has been checked
to work through several traivs builds, but it has been rebased, so CI
is disabled.