Added notice about implementing vector space using linear. Refs #211.
Bearriver uses the library simple-affine-spaces as an abstraction over
vector spaces. The library provides the necessary type-classes, but
users can still implement their own instances.

Because other vector libraries are much more common than
simple-affine-spaces, it is helpful to make users aware that using
Bearriver does not mean having to stick to that library, and they can
trivially define the necessary instances to use Bearriver constructs
like integral with their preferred vector implementations.

This commit adds a notice to the Bearriver module and to the integration
module, where users are most likely to benefit from that knowledge.