Add Functor, Applicative constraints in dunai-test. Refs #219.
The file dunai-test/src/FRP/Dunai/LTLPast.hs is missing Functor and
Applicative constraints that were assumed from the existence of a Monad
constraint in the same type signature. It does not compile in older
versions of GHC for that reason.

This change adds Functor and Applicative constraints to multiple
definitions in dunai-test, unconditionally (meaning that it is not
actually required in newer versions of GHC / base, which may trigger
a warning).

This change is the result of squashing commits that were checked
on Travis, so we are skipping the CI build.

[ci skip]