Add Functor constraint for backwards compatibility. Refs #219.
The file dunai-test/src/FRP/Dunai/LTLPast.hs is missing a Functor
constraint that was assumed from the existence of a Monad constraint in
the same type signature. It does not compile in older versions of GHC
for that reason.

This change as a functor constraint to the definition of untilSF,
unconditionally (meaning that it is not actually required in newer
versions of GHC / base).