Conditionally import fmap, <$>, <*> from applicative. Refs #218.
Dunai-test fails to compile with older versions of GHC because it
assumes (<$>), (<*>) and fmap are in the prelude. This change only came
about with base 4.8, and before it had to be imported from Data.Functor
or Control.Applicative.

This change adds a conditional import of those three from
Control.Applicative for versions older than 4.8.0.0.

This change has been checked through multiple builds on Travis, and
is a result of squashing commits, so we are skipping CI.

[ci skip]