Re-organize project into folders. Refs #220.
This project currently contains multiple libraries, many of
which are hidden in examples/ and extensions/ folders, all
of which makes discovery and maintenance harder.

This commit reorganizes the project into four main folders:
- Dunai
- Dunai Test
- Bearriver
- Examples.

The cabal.project file is kept at the top level and adapted to the new
folder locations.