Update installation instructions for haskanoid. Refs #220.
The performance section contains instructions to install haskanoid
with bearriver support.

These instructions are no longer valid with the re-org, but, also,
haskanoid is no longer included in this repo.

This commit modifies the instructions to install haskanoid straight
from the official repo, and use dunai and bearriver from hackage.

[ci skip]