Fix bug in parC. Refs #228.
The implementation of parC contains a bug that makes the output list
always have at most one element. This is caused by populating the list
of SFs to apply in parallel with just one SF irrespective of how many
elements are in the input signal (list).

This commit fixes that implementation by adding an extra operation that
delays the list initialization until the first input is received. From
that point on, the list can shrink, but not grow.