Update installation instructions in README. Refs #253.
The README suggests installation instructions that are not up to date with
current versions of Haskell development tools. In particular, they suggest
using `cabal sandbox`, a command that is no longer supported.

This commit modifies the README to provide up-to-date installation
instructions, tested with cabal-install 3.6.

It also modifies the installation command for haskanoid (the argument
haskanoid/ to the install command was incorrect).

Main author is @DireWolf707, fixes introduced by @ivanperez-keera.

[ci skip]