Update example in README. Refs #258.
The example in the README used an old dunai API and does not work with the
current version. In particular, the function embed_ won't be in scope with
the suggested import, the function liftS is now arrM, and the function
arrM_ is now constM.

This commit updates the example session to use the functions by their
new names, and includes substantially more detail.

[ci skip]