Fix implementation of dlSwitch. Refs #230.
The Yampa implementation includes additional features, among them an
implementation of branching SFs. These features are not normally
compiled with Yampa, but an separate branch allows users to enable those
features.

The module FRP.BearRiver.List implements the BearRiver counterpart of
branching SFs. The existing module had a number of errors that made it
invalid in Haskell. This commit reimplements the module and documents
it. It remains disconnected from the cabal file, to comply with the
Yampa implementation.

Original suggested fix submitted by @walseb.

Co-authored-by: Ivan Perez <ivan.perez@keera.co.uk>
Co-authored-by: Sebastian Wålinder <s.walinder@gmail.com>

[ci skip]