Remove commented code. Refs #261.
The module Control.Arrow.Util contains commented code.

Commented code, unless it's there to help understand a particular nuance
of the current implementation, should not be included in the
implementation itself. Instead, it should be either part of a tutorial,
readme, example, or be outside of the repo altogether.

This commit removes imports and definitions that are completely
commented out.