Add dSwitch to dunai. Refs #275.
The Control.Monad.Trans.MSF.Except module provides an immediate switch
operation, but not a decoupled or delayed switch function. This
limitation makes implementing decoupled switch harder, which affects
particularly the implementation of Bearriver's reactimate.

This commit introduces a separate dSwitch in
Control.Monad.Trans.MSF.Except that implements a delayed switching
function (the switch happens in the next sample).

Co-authored-by: Ivan Perez <ivan.perez@keera.co.uk>