Bearriver: fix error in reactimate when second input sensed is Nothing. Refs #266.
Reactimate takes as arguments two sensing actions: one to be used for
the first sample, and one to be used from that point on. If a sensing
action returns Nothing, the expectation is that the last value sensed
will be used.

In the current implementation, reactimate throws an undefined exception
if the continuing sensing action (second argument to reactimate) returns
Nothing on the first step.

The implementation of the continuing sensing action takes the value
received from the first sample, and holds it until a new value is
sensed. If that value is undefined, then the value returned by the
sensing action will also be undefined.

The problem stems from the fact that sense is implemented using
switching and passing an undefined as continuation for the signal
switched on to. Normally, the continuing sensing action would provide
inputs, so that continuation would not be used, but if the continuing
sensing action's first sensed input is Nothing, the continuation
received from switching will be used, which is, at that point,
undefined.

This commit simplifies the implementation by using instead a delayed
switching, and always providing the input sensed to the continuation.

Example code that caused error before this commit:

```Haskell
import Data.IORef
import FRP.Yampa

main = do
  ref <- newIORef (0.0::Double)
  reactimate senseI (senseRest ref) (actuate ref) sf

senseI = do
  putStrLn "SenseI"
  return 45

senseRest ref _ = do
  v <- readIORef ref
  putStrLn "SenseRest"
  return (0.1, Nothing)

actuate ref _ x = do
  v <- readIORef ref
  putStrLn $ "actuate " ++ show x
  writeIORef ref (v+1)
  return (v >= 1)

sf = identity
```

Co-authored-by: Ivan Perez <ivan.perez@keera.co.uk>