Use different integration rule in example. Refs #268.
The integral provided by Yampa and Bearriver uses the rectangle rule to
integrate. This quickly becomes inaccurate; a common sympton is that we
see unexpected loss or gains of energy in simulations that use gravity.

This commit modifies the example included with dunai to use the
trapezoid rule to integrate the velocity. Because it is linear, the
trapezoid rule will be accurate (up to floating-point accuracy), and not
present the symptoms described above.

Co-authored-by: ScottFreeCode <ScottFreeCode@users.noreply.github.com>
Co-authored-by: Miguel Negrao <miguel.negrao@friendlyvirus.org>