Remove unused, unexported module CausalCommutative. Refs #279.
The module Data.MonadicStreamFunction.CausalCommutative is not exported,
or imported by any module. Because it is dead code, it is best not to
keep it as part of the implementation.

The idea may still be useful and worth considering, but it should work
its way back into the implementation from the top (design-to-code).

This commit simply removes the module.