dunai: Use Data.Void type instead of the custom type Empty. Refs #79.
In Control.Monad.Trans.MSF.Except, we use our own uninhabited type
Empty. The Data.Void module, in base in versions of GHC after 7.10, and
in the package void prior to that, provides such a type.

Using a standard type makes the interface more intuitive for users
familiar with the type, provides features that our type may not provide
(e.g., instances, additional functions), delegates the implementation on
others, and reduces duplication.

This commit modifies the Control.Monad.Trans.MSF.Except module
to use the Void type instead of the Empty type.

The cabal file is modified to depend on the package void when using GHC
implementations up to 7.8.4.