dunai: Do not warn in the presence of orphan instances in Core. Refs #284.
Compilation of dunai reports a number of warnings in
Data.MonadicStreamFunction.Core related to orphan instances.
Having those instances separate from the type definition is a design
decision made in dunai in the name of extensibility. Therefore, we are
aware.

This commit disables the warning for having orphan instances in
Data.MonadicStreamFunction.Core.