dunai: Remove use of TupleSections. Refs #292.
The definition of step_ uses tuple sections to implement the function in
point-free style. The implementation is not clearer that way than naming
the argument, so the use of a language extension is not sufficiently
justified.

This commit implements step_ by naming the argument, removes that use of
TupleSections and adjusts pragmas.