dunai: Implement dot in MSF's instance of VectorSpace. Refs #298.
The MSF instance of VectorSpace does not give an implementation for dot.
The function is supposed to return a scalar, not an MSF of a scalar.

It was originally not possible to define such instances because the type
constraints on scalars Eq and Floating were not applicable to MSFs.
Version 0.2 of the library simple-affine-space remediated that situation
by moving the constraints to the default methods that use them, which
are not applicable here because we do not rely on the default
implementation.

This commit introduces:

- A change to the instance defition so that the type of the scalar is
  also an MSF.

- An implementation for dot.

- An implementation for multiplication and division that does not lift a
  scalar automatically and instead assumes it is an MSF.

- An implementation for normalize, since the existing implementation
  required Equality to be defined on scalars (which are now MSFs).

- Required additional constraints in the instance, namely Eq s, Num s,
  and Floating (MSF m a s).