dunai-test: Deprecate the function FRP.Dunai.LTLPast.prev. Refs #312.
The function prev, which delays an MSF by one sample, is needed in the
module, but should not be exposed. It is also not currently used
anywhere in dunai-test. Consequently, prev should be hidden from the
interface.

This commit deprecates the top-level prev to signal to users that it
will be removed in the future.