bearriver: Add NFData instance for Event. Refs #326.
Add NFData instance for Event, present in Yampa. This instance is
required to deep seq values, which is necessary especially when a game
state contains an event.

This commit requires a new dependency in bearriver's Cabal file. We
include version 1.3.0.0 of deepseq because it is compatible and included
in older versions of GHC.