bearriver: Update Copyright notices. Refs #332.
Bearriver borrows code from Yampa, yet the relation between the two is
not made clear in the license.

This commit updates the copyright owners and the license to reflect the
main contributors over the years, and also to indicate that the project
borrows heavily from Yampa.

The attribution of Yampa, the agreement on the years in the notices, and
the people listed, have been agreed to with the respective authors,
including Henrik Nilsson, Antony Courtney, George Giorgidze and Manuel
Baerenz.