dunai-test: Remove FRP.Dunai.LTLPast.prev. Refs #341.
The function FRP.Dunai.LTLPast.prev was deprecated in dunai-test-0.9. As
it's now been 3 versions since it was deprecated, it can now be removed
from the public interface.

This commit inlines prev in the definition of prev', as it was being
used by it. I choose to inline it instead of keeping it as a local
definition because it is small, and the approach is consistent with
other definitions in the same file (history', ever').