dunai: Update paper links to ACM, personal website. Refs #346.
Both the cabal file and the dunai's README contain links to a mirror for
the paper "Functional Reactive Programming, Refactored".

Those links point to the old website at UoN, but my new website is
https://ivanperez.io/. The links should be adjusted accordingly.

This commit updates all pointers to my UoN website in the README to now
point to my new website. It also updates the links to ACM after
re-authorizing the article links, and adds links to portions of my
personal website that are authorized to link to my papers.