dunai: Update paper links to ACM, personal website. Refs #346.
Both the cabal file and the dunai's README contain links to a mirror for
the paper "Functional Reactive Programming, Refactored".

Those links point to the old website at UoN, but my new website is
https://ivanperez.io/. The links should be adjusted accordingly.

This commit updates the cabal file to point to an anchor in my personal
website, from which the FRP Refactored paper can be obtained.