bearriver: Align signature consistently (KSG Haskell 1.3.0 - 5.5). Refs #347.
Prefer consistent names, structure, and improvements over localized
optimiza- tions. For example, a module might define an API in a
particular order, and a module with a similar but distinct API might
list elements in a completely different order. Whenever possi- ble, code
should be consistent.

This commit aligns a signature consistenly with other signatures and
other elements in the same signature.