bearriver: Remove vertical spacing (KSG Haskell 1.3.0 - 5.5). Refs #347.
Prefer consistent names, structure, and improvements over localized
optimiza- tions. For example, a module might define an API in a
particular order, and a module with a similar but distinct API might
list elements in a completely different order. Whenever possi- ble, code
should be consistent.

This commit removes an extra spacing in a haddock comment for
consistency with other haddock comments.