dunai-test: List exports explicitly (KSG Haskell 3.4). Refs #348.
The module declaration should explicitly list the definitions exported
by the module.

The module declaration should use haddock indicators to group the
elements exported so as to facilitate navigating the documentation.

We draw inspiration from yampa-test to select the elements exported.
This restrics the API and therefore carries with it a major version
bump.