dunai-test: Wrap lines to 80 characters (KSG Haskell 1.3.0 - 4.3). Refs #348.
We set the code limit for Haskell files at 80 characters. A "character"
means any Unicode code point. Any line that would exceed this limit
should be line-wrapped.

This commit simplifies part of the code. The original code was going
over the column limit, and I decided that it was easier to simplify it
than to try to awkwardly make it fit in 80 columns, which was resulting
in substantial changes in an effort to keep consistency across the
multiple cases of a function.