bearriver: Port module FRP.Yampa.Basic to bearriver. Refs #376.
The module FRP.Yampa.Basic is being offered by Yampa, but not bearriver.
This makes the interface offered by bearriver not compliant with Yampa,
therefore not delivering on the promise of being API-compatible.

This commit adds the module FRP.Yampa.Basic into bearriver, as
FRP.BearRiver.Basic. Several definitions that existed in FRP.BearRiver
but belong in this new Basic module are moved accordingly and
re-exported as needed. The auxiliary definition of replaceOnce is no
longer needed, so it is removed.