bearriver: Port module FRP.Yampa.Event to bearriver. Refs #380.
The module FRP.Yampa.Event is being offered by Yampa, but not bearriver.
This makes the interface offered by bearriver not compliant with Yampa,
therefore not delivering on the promise of being API-compatible.

This commit adds the module FRP.Yampa.Event into bearriver, as
FRP.BearRiver.Event. Several definitions that existed in FRP.BearRiver
but belong in this new Event module are moved accordingly and
re-exported as needed.

The module is also structured and commented similar to Yampa's
counterpart, to minimize the differences between the two and reduce the
maintenance burden.

Some definitions pertaining to Events, like those to convert from types
Bool, and those to convert Event to Maybe, are not offered by Yampa, so
they are left in the BearRiver module although they might conceptually
fit best in a module dealing with Events. Those functions will have to
be made private, be removed, or be introduced in Yampa before they are
introduced in BearRiver.