dunai: Increase max-backjumps in CI setup. Refs #381.
Some of the examples included with dunai and bearriver use SDL-gfx.
Specifically, the package classicfrp uses SDL-gfx < 0.7, which is not
the latest version available.

For some reason, cabal is getting confused and not trying 0.6.2, which
would be compatible with all packages to be installed/tested. If that
version constraint is increased, then cabal gets confused about the
versions of hashable and aeson, dependencies of criterion.

This commit adds `--max-backjumps=10000" to the installation and testing
commands in the CI script, which is enough for cabal to find suitable
installation plans with all versions of GHC supported by dunai.