bearriver: Port module FRP.Yampa.Scan to bearriver. Refs #383.
The module FRP.Yampa.Scan is being offered by Yampa, but not bearriver.
This makes the interface offered by bearriver not compliant with Yampa,
therefore not delivering on the promise of being API-compatible.

This commit adds the module FRP.Yampa.Scan into bearriver, as
FRP.BearRiver.Scan. Definitions are moved from the main module
FRP.BearRiver to the new module FRP.BearRiver.Scan. One of the
definitions was clearly undefined, so we copy the definition from Yampa.