bearriver: Port module FRP.Yampa.Delays to bearriver. Refs #384.
The module FRP.Yampa.Delays is being offered by Yampa, but not
bearriver. This makes the interface offered by bearriver not compliant
with Yampa, therefore not delivering on the promise of being
API-compatible.

This commit adds the module FRP.Yampa.Delays into bearriver, as
FRP.BearRiver.Delays. Definitions are moved from the main module
FRP.BearRiver to the new module FRP.BearRiver.Delays.