dunai: Document GHC 9.8.1 support in readme. Refs #387.
The README states that the library is compatible with GHC up to 9.4, but
it's compatible with GHC 9.8. Also, there's no need to recommend a
version of GHC lower than 9: it's now quite old and may give the wrong
impression.

This commit updates the line stating the compatibility of dunai to
include GHC 9.8.1, and removes the line recommending older versions of
GHC or cabal.