dunai: Mark package as uncurated. Refs #394.
Currently, hackage maintainers could publish a revision to our cabal
file if something needs fixing. This creates an issue from the point of
view of software maintenance, since the change is neither approved by
us, nor is it propagated back to us.

This commit modifies the cabal file to indicate that the package should
not be modified by Hackage maintainers.