dunai: Implement List interface using list-transformer. Refs #418.
The Control.Monad.Trans.MSF.List module relies on an implementation of
ListT that is outdated and has been removed from newer versions of the
transformers library.

This commit introduces a variant of the module implemented using the
ListT definition from list-transformer.

Picking implementation is controlled via a Cabal flag, which is by
default set to the old implementation, but can be automatically set to
the new one so that Cabal picks it up automatically when compiling with
versions of transformers that do not provide ListT anymore.

Deprecation notices are attached to the functions that use the old
definition of ListT.