bearriver: Offer all definitions from FRP.Yampa.Integration. Refs #422.
The module FRP.BearRiver.Integration does not contain all definitions offered
by its Yampa counterpart. This makes the interface offered by bearriver
not compliant with Yampa, therefore not delivering on the promise of
being API-compatible.

This commit adds missing definitions to the module FRP.BearRiver.Integration.
Comments are changed inside the module to match FRP.Yampa.Integration.

A definition is hidden in an import in FRP.BearRiver, since it now
clashes with one being offered by this module.