bearriver: Remove occassionally due to name clash. Refs #423.
The introduction of a Random module in bearriver carries with it an
occassionally function with a slightly different signature than the
current function. This creates an incompatibility that just cannot be
resolved via hiding and re-exporting.

This commit removes the existing occassionally function from the module
FRP.BearRiver, and a dependency that no longer applies from the cabal
file.