dunai: Increase upper bounds on transformers. Refs #427.
Support for transformers-0.6 in dunai 0.13 does not yet work correctly.
Dunai unconditionally depends on transformers < 0.6 and, if the
list-transformer flag is set, additionally depends on transformers <
0.7. The constraints are conjunctive, so dunai always has an upper
bound of 0.6 on transformers. This prevents dunai from installing with
transformers-0.6.

This commit adjusts the rules involving transformers so that they are
always governed by the list-transformer flag.