dunai-examples: Increase upper bounds on mtl, transformer. Refs #427.
It seems the transformers-0.6 support in dunai 0.13 (following
discussion #368) does not yet work correctly. The mtl < 2.3 constraint
on in the examples and the ==0.5.* both exclude transformers-0.6.

This commit adjusts the example's cabal file to allow installation with
mtl versions lower than 2.4, and transformer lower than 0.7, which
together are enough to allow installation with transformers-0.6.