bearriver: Document module. Refs #426.
BearRiver does not contain in FRP.BearRiver.Switches the documentation
included at the beginning of FRP.Yampa.Switches. This makes the offer of
bearriver not as good as Yampa's, failing to deliver on the promise of
being a suitable Yampa replacement.

This commit adds the module documentation to FRP.BearRiver.Switches.