bearriver: Publish embed specialized for FRP.Yampa's SF. Refs #439.
The function embed doesn't have the same signature as in Yampa: in
bearriver it returns the result in a monadic context, but in Yampa it
just returns a list of values. This makes modules that use Yampa's embed
not API-compatible with bearriver.

This commit hides bearriver's embed in the FRP.Yampa module published by
bearriver, and instead publishes a specialized version with the same
signature as Yampa's, by assuming that the monad is the Identity monad.