dunai-examples: Add module documentation (KSG Haskell 1.3.0 - 3.3). Refs #409.
Rule: All Haskell files must include a copyright notice in a comment.
Additional notices at the top may include the full Copyright license, a
pointer to the license contained in a file, or other conﬁdentiality
notices and warnings.

This commit adds a missing copyright notice, as well as haddock
documentation stating the purpose of the module.