dunai-examples: Fix bug in example where screen hangs on click. Refs #457.
The example dunai-examples/list/ is supposed to demonstrate how new
elements can be added dynamically in a dunai animation by leveraging the
list monad, by allowing users to create new elements with a mouse click.
However, it currently hangs when one clicks on the screen.

Specifically, a switch is currently stuck in an infinite loop.

This commit introduces a `notYet` that delays the switch event to the
next point in time T(n + 1). The switch will now happen AFTER the event
at T(n) has passed.

This issue is described in the [Yampa Arcade
paper](https://dl.acm.org/doi/10.1145/871895.871897)
on page 8.

> The resulting signal function is composed with
> `notYet :: SF (Event a) (Event a)`
> that suppresses initial event occurrences.
> Thus the overall result is a source of kill and
> spawn events that will not have
> any occurrence at the point in time when it is
> first activated. This is to prevent gameCore
> from getting stuck in an infinite loop of
> switching.