bearriver: Implement reactimate function matching the signature in Yampa. Refs #454.
The function reactimate doesn't have the same signature as in Yampa: in
bearriver it forces the monad of the SF to be the same as the monad of
the sensing/actimation actions and reactimate itself, but a Yampa SF is
mapped to a Bearriver SF on the Identity monad. This makes modules that
use Yampa's reactimate not API-compatible with bearriver.

This commit hides FRP.BearRiver.reactimate in Bearriver's FRP.Yampa
module, and defines an alternative that has the same signature as
Yampa's.