dunai: Remove deprecated functions related to list-t. Refs #445.
dunai includes several deprecated definitions that were deprecated three
versions ago when support for using list-transformer (instead of list-t)
was introduced. As per our policy of waiting three versions since
deprecation until a function is removed, these definitions can be safely
removed.

This commit removes the deprecated functions, changes the GHC options
and all conditionals related to picking different versions of ListT, and
all related imports. The documentation is adjusted to remove the mention
on the list monad having constraints on the monad it is applied to.